Cleaned up types.

+++ src/components/ogre/authoring/IPolygonPositionProvider.h
@@ -48,7 +48,7 @@ struct IPolygonPositionProvider
 	 * @param localPosition The local position within the polygon's space.
 	 * @return The height.
 	 */
-	virtual float getHeightForPosition(const WFMath::Point<2>& localPosition) const = 0;
+	virtual double getHeightForPosition(const WFMath::Point<2>& localPosition) const = 0;
 
 };
